+++ crates/parser/src/lib.rs
@@ -1,5 +1,6 @@
 mod bibtex;
 mod build_log;
+mod config;
 mod latex;
 
-pub use self::{bibtex::parse_bibtex, build_log::parse_build_log, latex::parse_latex};
+pub use self::{bibtex::parse_bibtex, build_log::parse_build_log, config::*, latex::parse_latex};