+++ CHANGELOG.md
@@ -11,9 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Allow optionally passing cursor position to `textDocument/build` request for use in forward search after building.
   Previously, the server had to guess the cursor position ([#475](https://github.com/latex-lsp/texlab/issues/475))
-
-### Changed
-
+- Add experimental `texlab.experimental.citationCommands` setting to allow extending the list of citation commands
+  ([#832](https://github.com/latex-lsp/texlab/issues/832))
 - Add support for escaping placeholders in build arguments similar to forward search
 
 ### Fixed